kite: ignore extra o frames
This is a guess - did not manage yet to repro browser <-> kloud
connection problems to make it possible to inspect req/resp.

This reverts a change in kite that may be causing disconnections,
when sockjs client does send multiple o frames.

Going to push to upstream if it fixes the problem.