stackplan: fix BuildMachines
Fixes erroneous validation logic, which reported error if a
jMachine.Users had a user without owner right.

The purpose of such check is to ensure that apply action is only
performed by an owner, never by a user with insufficient permissions.

However that check exists in a different part of the Builder logic,
and the check removed here is simply wrong - it causes that any
jMachine that was shared with different user is going to be reported
as invalid.