klientctl: healthcheck of failed start
Klient can "fail to start" even if the users internet is down, or
other external issues. By health checking, we attempt to give them
accurate information of possible causes for klient not starting.

Also, for debugging, we health check before even starting klient - just
to have a record of if the user tried to start klient before their
internet was fully up *(mainly for debugging purposes)*.