client/home: fix React 15 problem for ToggleButton
It was giving an error saying you are changing an uncontrolled input into a controlled one... This is due to passing `undefined` for required valus from Inputs. In this case `checked` was `undefined` at first, and then it's being toggled to `true` or `false`. It being `undefined` at first makes it an uncontrolled input, while passing `false|true` for `checked` it becomes an controlled input.

To make this work correctly we never pass `undefined` now.

Signed-off-by: Umut Sirin <umutsirin1@gmail.com>