go/presence: set non active sub-ed teams presence as processed
store non active sub-ed team's ping requests
as processed? We wont be charging trailing teams before second month's
payment is due, so we start collecting presence info( with processed
false )  after the first month, and first month is completely free.
Second issue is, when a sub is in non-active state, we should still
collect presence info but we wont be charging users during that period,
because we dont allow them to utilize koding