Bongo: handleRequest: make sure callback of a given request called once
There is no check for preventing to call a given method callback's more
than once. This causing some unexpected issues on the application. And
it's preventing us to use async.parallel since it's not allowing this
situation. To prevent that we are keeping the unique callbackId of the
call and checking it's existence before taking any action. Also for
debug purposes this change provides logs with following format;

 - [Bongo:LEAKED_CALL] CONSTRUCTUR_NAME.METHOD_NAME