client/mocks: add mock managed machine
This is the main machine object which is used to represent a test
machine. Main reason we are using a managed machine is because a managed
machine doesn't require a stack/stack template to be built. Our frontend
already knows about this, and doesn't expect managed machines to have a
stack/template so that it immediately shows the machine.

We want to be able to test IDE features such as shortcuts, so having a
managed machine trims out the requirements of other entities.

We will create a MemoryFs instance, and a MemoryKlient and this
MemoryKlient will always return the mock managed machine to make our
frontend think that this is a real machine.