object: rework Walk -> Replace
In order to implement both traversing over all fields of value
and ensure all fields could be set / replaced, one would
need much more complex implementation of the Walk method.

For our purpose the Walk would be used to implement
a function which will replace ShadowVariables - it will
walk over a JSON's unmarshalled object and replace
all secrets in string fields.

So instead of making Walk work, this PR reworks it
for the particular purpose of replacing string keys.

Simpler implementation and does the exact job.