Bongo: loadConstructors becomes loadModels with caching feature
Currently we are requiring all the files under provided models directory
or directories, with this change Bongo will try to read models.json
which looks like the one provided with this commit. And it will load
the defined models underneath. If this file not found, Bongo will walk
in directories as before and load models also it will create the missing
models.json under provided models path or paths.

After this one in, if one wants to add a new model, s/he needs to update
models.json or needs to remove and restart server to auto-generate new
models.json file. Current one includes all the existing models and auto-
generated by Bongo already.