PermissionSet: permit: with result also pass the passed permission index
In the case of multiple/advanced permissions we need to know which one
passed and allowed by permit in the target method context. So, while
checking out the advanced permissions we're sending back the passed one
via client object. In any method we can get this information by looking
`client._allowedPermissionIndex` field which will give the index of
passed permission.

So, if you write something like this;

  foo: permit

    advanced: [
      { permission   : 'some permission', validateWith: Validators.own }
      {
        permission   : 'access related'
        validateWith : accessValidator ACCESSLEVEL.READ
      }
      { permission   : 'modify object', superadmin: yes }
    ]

    success: (client, callback) ->
      ...

and this permission is granted because you were a superadmin then
client._allowedPermissionIndex will become 2. But if you are owner of
this instance this time it will become 0 because the first rule granted.