kd: use default team for "credential create"
TODO It should be possible to create credentials
which are global (no need for team name), however
socialapi requires team name in order to authenticate
(clientId obtained by username/team lookup). That's why
the team name is still there.

We should silently attach active team name so kloud
can use it to insert credential to sneaker. And
set visibility to global for that credential afterwards
(via relationship to jAccount document).