scripts/api-generator: make it possible to read error object
The remote.api endpoint on 200 OK returns either response object
if request was successful, or returns an error object to indicate
an error.

This makes it impossible to inspec error object for methods that
don't use DefaultResponse schema.

This PR makes it possible to read both - response object or
eventual error.