kd: do not fetch kloud kite from Kontrol
The code responsible for the kontrol fallback was
written before kd started to use single baseurl
for all the services. Also it was not possible
to overwrite the configuration at the point,
so it was infact a workaround for incomplete
configuration.

Since then "kd config set" and "kd config reset"
were introduced to ensure proper configuration.

Additionally there's a security risk involved
with the current implementation: there is
nothing to prevent a malicious user to use
KD's kite.key to register a malicious kite
named "kloud" to the production Kontrol.

This way the user could e.g. steal passwords
by logging auth.passwordLogin requests.