cmd/loopfuse: FUSE-backed mount-like tool for testing/debugging
This tool implements "mount -o rbind". The following command:

    $ loopfuse /tmp/source /tmp/dest

Bind-mounts /tmp/source directory to a /tmp/dest one. It creates
a cache in the /tmp directory, to fake how real rsync-based
implementation would work.

The -v flag turns on FUSE debug output.

    $ loopfuse -v /tmp/source /tmp/dest

The -tmp flag can be used to persist cache directory for later
inspection or to start a mount with a dirty cache:

    $ loopfuse -v -tmp /tmp/cache /tmp/source /tmp/dest