(wip) klient/machine: add exec/kill handlers
This PR adds the following kite methods:

  - os.exec
  - os.kill
  - machine.exec
  - machine.kill

The os.exec methods is a reworked version of existing
exec method with a new API that is needed for KD. Since
the exec method is still used on client side, it's being
marked as deprecated and new version rolled out.

The os.kill method addresses shortcoming in current "kd run"
implementation - if user aborts command execution with CTRL+C,
the command won't get killed on a remote machine.

New implementation is going to call machine.kill upon exit,
which effectively will kill the user process on the remote side.

The os.kill method is limited to only those processes that
were started with os.exec.