klient: remove wrong timeout
The intention was to stop polling when remote end went away.
The right way to implement this is to have sockjs heartbeats
in client, the HTTP timeout is used for roundtrip timeout,
when it kicks in it abruptly aborts HTTP connection, which
is what we do not want here, since the connection may
be just idle.