kloud/metadata: add CloudInit type with Merge function
The Merge function is going to be used to merge kloud's
cloud-init, which will be responsible for deploying
klient service with a customized provision.sh, into
user's cloud-init content.

After metadata package is ready, kloud is going
to be refactored to not depend on S3 bucket lookup
to install a klient, thus S3 credential configuration
can be removed from kloud, making it effectively
do not depend directly on S3 credentials anymore.

After the above is in place, kloud/userdata package
is going to be removed.