kloud/provider: remove user_data interpolation
The user_data interpolation was used in previous
implementation to workaround a terraform interpolation
restriction, which caused stack build to fail when
user_data contained variables.

It is not needed anymore as user_data handling was
reworked - we merge user_data with kloud's cloud-init
and interpolate it at once.

Moreover it broke count=2+ builds, as count.index
was always 1 within the null_resource scope - so
all the instances used the same kite key.

Fixes #10967.