kites/config: rename local -> mount
This PR renames configuration from local to mount. The purpose
is to enable for mount configuration settings and not have
two places where mounts are configured.

So the following:

    $ kd config set local.mountHome ~/mount

Changes to:

    $ kd config set mount.home ~/mount

And the following:

    $ kd config set local.mounts.default ~/mount

Changes to:

    $ kd config set mount.exports.default ~/mount

No backward-compatiblity is ensured, as the mount
configuration is not widely deployed. If needed,
it is possible to add backward-compatibility later on.