kd: rework bolt handling
This PR reworks handling of kd.*.bolt database file in respect of writes
and reads, so it is possible to execute a kd processes concurrently
with another long-running kd process.

This PR makes the database access more fine grained - during init
(and throught the process lifetime) a read-only access is requested,
which allows for multiple processes to start concurrently.

Upon exit a read-write access is acquired, which makes kd processes,
which exit at the same time, queue with a 3s timeout in order to flush
their data to a database.

Closes #11154.