kloud/stack: add userdata to provider schema
A "userdata" is a JSONPath which allows for locating
a user_data (cloud-init) key within a stack template tree.

Locating cloud-init in the tree is going to allow for
provider-agnostic manipulation of stack template content,
like creating new and custom templates on-the-fly (as oppose
to dummy templates received from JTemplate.Sample).

The templates are meant to be manipulated client-side (e.g. kd),
before sending them over to koding.